feat: add /vouchers endpoint
We have a way to GET a voucher by GUID but there's no way to know the
vouchers on the server. We need _at least_ an API like this to list the
GUIDs it knows about. We definitely need more fields exposed but this
will get us started and it's closer to having something which resembles
real life in production at least.

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>