fix: packit configuration
Make sure the %global commit variable in the spec
is replaced with the actual commit hash.

Replace the Version: and Release: tags with the ones expected
by Packit.

Replace create-archive pre-sync and post-upstream-clone with
post-modifications which is common to the release sync actions.
See [1]

Add askalomo-cli missing dependencies.

Do not install golang 1.25.0 with gvm use GOTOOLCHAIN=auto instead

Remove not needed Makefile target related to packit

[1] https://packit.dev/docs/configuration/actions#release-synchronization-actions

Signed-off-by: Miguel Martín <mmartinv@redhat.com>