fix: download source tarball in post-modifications action
The post-modifications action was failing in the propose_downstream
workflow because go_vendor_archive requires the source tarball (Source0)
to exist before it can create the vendor archive.

In the Packit sandcastle build environment, the source tarball is not
available at the expected location when post-modifications runs. This
caused the error:

  FileNotFoundError: [Errno 2] No such file or directory:
  '.../go-fdo-server-<commit>.tar.gz'

Add spectool command to download the source tarball from the GitHub
release before running go_vendor_archive. Also add rpmdevtools to
srpm_build_deps to ensure spectool is available in the build environment.

This fix allows the propose_downstream job to succeed when triggered by
a new GitHub release.

Fixes: https://dashboard.packit.dev/jobs/propose-downstream/22429

Signed-off-by: Paul Whalen <pwhalen@fedoraproject.org>